requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx>=0.27.0,<0.28.0
pydantic==2.4.2
python-dotenv==1.0.0
ollama==0.3.3
mysql-connector-python==8.1.0
sqlalchemy==2.0.23
mysql-connector-python==8.1.0
sqlalchemy==2.0.23